Refactor import paths to use app_config instead of config for consistency across modules
