+++ libs/bootstrap/configuration.py
@@ -17,7 +17,7 @@ import libs.commands.report.entry
 import libs.commands.results.entry
 import libs.global_value as g
 from integrations import factory
-from libs.bootstrap.config import AppConfig
+from libs.bootstrap.app_config import AppConfig
 from libs.commands.registry import member, team
 from libs.data import initialization, lookup
 from libs.functions import compose
