+++ tests/database/conftest.py
@@ -12,7 +12,7 @@ import pytest
 
 import libs.global_value as g
 from libs.bootstrap import configuration
-from libs.bootstrap.config import AppConfig
+from libs.bootstrap.app_config import AppConfig
 from libs.data import initialization, lookup
 from libs.utils import dbutil
